requirements.txt
python-telegram-bot>=20.7
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
e2b>=2.6.3
httpx>=0.27.0
httpx==0.25.2
asyncio-throttle==1.0.2
redis==5.0.1
docker==6.1.3
rich==13.7.0
typer==0.9.0
psutil==5.9.6
requests==2.31.0
aiofiles==23.2.1